Rename career-hero to hero-career to follow hero block guidelines (#78)
* Rename career-hero to hero-career to follow hero block guidelines

This also fixes a layout issue with this block
